backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
langchain==0.0.267
openai==0.28.0
python-dotenv==1.0.0